Refact: Remove usages of force unwrap, Remove usages of observe all state in WithViewStore

+++ Blinterns_ToDo_App/Domain/TabBar/Domain/ToDoList/ToDoListReducer.swift
@@ -10,6 +10,8 @@ import ComposableArchitecture
 
 extension ToDoListReducer {
     
+    //TODO: Force Unwrap is a no no
+    
     @ReducerBuilder<State, Action>
     var core: some ReducerProtocol<State, Action> {
         Reduce { state, action in
@@ -22,14 +24,17 @@ extension ToDoListReducer {
                     state.toDoFormState = .init(category: state.filterValue)
                     return .none
                 case .onDeleteButtonTapped:
-                    for id in state.finishedToDoIdList[state.filterValue]! {
+                    guard let observedFinishedToDoId = state.finishedToDoIdList[state.filterValue] else {
+                        return .none
+                    }
+                    for id in observedFinishedToDoId {
                         do {
                             try toDoRepository.deleteToDo(id)
-                            if let toDo = state.toDoList[.all]!.first(where: { $0.id == id }) {
-                                state.toDoList[.all]!.remove(toDo)
-                                state.toDoList[toDo.category]!.remove(toDo)
-                                state.finishedToDoIdList[.all]!.removeAll(where: { $0 == toDo.id })
-                                state.finishedToDoIdList[toDo.category]!.removeAll(where: { $0 == toDo.id })
+                            if let toDo = state.toDoList[.all]?.first(where: { $0.id == id }) {
+                                state.toDoList[.all]?.remove(toDo)
+                                state.toDoList[toDo.category]?.remove(toDo)
+                                state.finishedToDoIdList[.all]?.removeAll(where: { $0 == toDo.id })
+                                state.finishedToDoIdList[toDo.category]?.removeAll(where: { $0 == toDo.id })
                             }
                         } catch {
                             state.alertState = .init(title: {
@@ -67,24 +72,24 @@ extension ToDoListReducer {
             case .internal(let action):
                 switch action {
                 case .handleFinishedListFilter(let toDo):
-                    if let allIndex = state.toDoList[.all]!.firstIndex(where: { $0.id == toDo.id }) {
+                    if let allIndex = state.toDoList[.all]?.firstIndex(where: { $0.id == toDo.id }) {
                         
-                        state.toDoList[.all]![allIndex] = toDo
+                        state.toDoList[.all]?[allIndex] = toDo
                         
-                        if let allFinishedIndex = state.finishedToDoIdList[.all]!.firstIndex(where: { $0.self == toDo.id }) {
-                            state.finishedToDoIdList[.all]!.remove(at: allFinishedIndex)
+                        if let allFinishedIndex = state.finishedToDoIdList[.all]?.firstIndex(where: { $0.self == toDo.id }) {
+                            state.finishedToDoIdList[.all]?.remove(at: allFinishedIndex)
                         } else {
-                            state.finishedToDoIdList[.all]!.append(toDo.id)
+                            state.finishedToDoIdList[.all]?.append(toDo.id)
                         }
                         
-                        if let targetIndex = state.toDoList[toDo.category]!.firstIndex(where: { $0.id == toDo.id }) {
+                        if let targetIndex = state.toDoList[toDo.category]?.firstIndex(where: { $0.id == toDo.id }) {
                             
-                            state.toDoList[toDo.category]![targetIndex] = toDo
+                            state.toDoList[toDo.category]?[targetIndex] = toDo
                             
-                            if let targetFinishedIndex = state.finishedToDoIdList[toDo.category]!.firstIndex(where: { $0.self == toDo.id }) {
-                                state.finishedToDoIdList[toDo.category]!.remove(at: targetFinishedIndex)
+                            if let targetFinishedIndex = state.finishedToDoIdList[toDo.category]?.firstIndex(where: { $0.self == toDo.id }) {
+                                state.finishedToDoIdList[toDo.category]?.remove(at: targetFinishedIndex)
                             } else {
-                                state.finishedToDoIdList[toDo.category]!.append(toDo.id)
+                                state.finishedToDoIdList[toDo.category]?.append(toDo.id)
                             }
                             
                         }
@@ -103,8 +108,8 @@ extension ToDoListReducer {
                     case .external(let action):
                         switch action {
                         case .onToDoAdded(let newToDo):
-                            state.toDoList[newToDo.category]!.append(newToDo)
-                            state.toDoList[.all]!.append(newToDo)
+                            state.toDoList[newToDo.category]?.append(newToDo)
+                            state.toDoList[.all]?.append(newToDo)
                             return .send(.toDoFormAction(.dismiss))
                         default:
                             return .none
@@ -123,18 +128,18 @@ extension ToDoListReducer {
                     case .external(let action):
                         switch action {
                         case .onToDoDeleted(let toDo):
-                            state.toDoList[.all]!.remove(toDo)
-                            state.toDoList[toDo.category]!.remove(toDo)
-                            state.finishedToDoIdList[.all]!.removeAll(where: { $0 == toDo.id })
-                            state.finishedToDoIdList[toDo.category]!.removeAll(where: { $0 == toDo.id })
+                            state.toDoList[.all]?.remove(toDo)
+                            state.toDoList[toDo.category]?.remove(toDo)
+                            state.finishedToDoIdList[.all]?.removeAll(where: { $0 == toDo.id })
+                            state.finishedToDoIdList[toDo.category]?.removeAll(where: { $0 == toDo.id })
                             return .send(.toDoDetailAction(.dismiss))
                         case .onToDoIsFinishedToggled(let toDo):
                             return .send(.internal(.handleFinishedListFilter(toDo)))
                         case .onToDoUpdated(let toDo):
-                            if let targetIndex = state.toDoList[.all]!.firstIndex(where: { $0.id == toDo.id }) {
-                                state.toDoList[.all]![targetIndex] = toDo
-                                if let targetIndex = state.toDoList[toDo.category]!.firstIndex(where: { $0.id == toDo.id }) {
-                                    state.toDoList[toDo.category]![targetIndex] = toDo
+                            if let targetIndex = state.toDoList[.all]?.firstIndex(where: { $0.id == toDo.id }) {
+                                state.toDoList[.all]?[targetIndex] = toDo
+                                if let targetIndex = state.toDoList[toDo.category]?.firstIndex(where: { $0.id == toDo.id }) {
+                                    state.toDoList[toDo.category]?[targetIndex] = toDo
                                 }
                             }
                             return .none
